BackEnd_New-main/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.103
botocore==1.34.103
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect2==0.7.0
Django==4.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
fonttools==4.51.0
frozenlist==1.4.1
html5lib==1.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
openai==0.28.0
pdfkit==1.0.0
pillow==10.3.0
pycparser==2.22
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tqdm==4.66.2
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
weasyprint==61.2
webencodings==0.5.1
yarl==1.9.4
zopfli==0.2.3
